test: split toSafeTransferRows specs out of assemble.spec to satisfy max-lines
